Moving cache_ctx.go to the keeper
