+++ modules/async-icq/keeper/relay.go
@@ -1,10 +1,8 @@
 package keeper
 
 import (
-	"github.com/cosmos/ibc-apps/modules/async-icq/v7/types"
-	"github.com/cosmos/ibc-apps/modules/async-icq/v7/utils"
-
 	"cosmossdk.io/errors"
+	"github.com/cosmos/ibc-apps/modules/async-icq/v7/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -31,7 +29,7 @@ func (k Keeper) OnRecvPacket(ctx sdk.Context, packet channeltypes.Packet) ([]byt
 
 	// If we panic when executing a query it should be returned as an error.
 	var response []byte
-	err = utils.ApplyFuncIfNoError(ctx, func(ctx sdk.Context) error {
+	err = applyFuncIfNoError(ctx, func(ctx sdk.Context) error {
 		response, err = k.executeQuery(ctx, reqs)
 		return err
 	})
